temp save still need to find error

+++ tetrahedron/mesh_struct/graph_color.cpp
@@ -27,7 +27,7 @@ void GraphColor::graphColorTet(MeshStruct& mesh_struct, int different_color_stra
 
 	std::vector<int>element_order(mesh_struct.tet_tet_index.size());
 
-	int max_color_num = max_degree / 3;
+	int max_color_num =(int) ((double) max_degree) / 3;
 
 	for (int i =0; i < element_order.size(); ++i) {
 		element_order[i] = i;
